introduce solidity_testing_utils crate
No functional changes. Split the cleanup/refactoring of testing crates into a separate PR before introducing new ones later in the stack:

- `codegen_testing` crate:
    - remove unused deps.
    - add `versions` module and make it public.
- add `solidity_testing_utils`:
    - for helpers used through multiple testing crates.
    - make it own serializing snapshots.
    - moved snapshots under it.
