+++ code-analysis/Cargo.toml
@@ -11,6 +11,7 @@ members = [
     "crates/solidity/outputs/rust/lib",
     "crates/solidity/outputs/rust/tests",
     "crates/solidity/outputs/typescript/lib",
+    "crates/solidity/testing/utils",
 ]
 
 [workspace.package]
@@ -46,6 +47,7 @@ serde_json = { version = "1.0.81" }
 serde_yaml = { version = "0.9.14" }
 similar-asserts = { version = "1.4.2" }
 solidity_rust_lib = { path = "crates/solidity/outputs/rust/lib" }
+solidity_testing_utils = { path = "crates/solidity/testing/utils" }
 typed-arena = { version = "2.0.1" }
 unicode_names2 = { version = "0.6.0" }
 walkdir = { version = "2.3.2" }
