+++ code-analysis/crates/codegen/testing/Cargo.toml
@@ -10,4 +10,3 @@ codegen_schema = { workspace = true }
 codegen_utils = { workspace = true }
 semver = { workspace = true }
 walkdir = { workspace = true }
-Inflector = { workspace = true }
