+++ code-analysis/crates/solidity/testing/snapshots/cst_output/ContractDefinition/empty_contract/input.sol
