+++ code-analysis/crates/solidity/testing/snapshots/cst_output/ContractDefinition/missing_field_type/input.sol
