+++ code-analysis/crates/solidity/testing/snapshots/cst_output/ContractDefinition/unterminated_body/input.sol
