+++ code-analysis/crates/solidity/testing/snapshots/cst_output/ExponentiationExpression/operator_associativity/input.sol
