+++ code-analysis/crates/solidity/testing/snapshots/cst_output/InterfaceDefinition/sample_counter/input.sol
