+++ code-analysis/crates/solidity/testing/snapshots/cst_output/MultilineComment/multiple_lines/input.sol
