+++ code-analysis/crates/solidity/testing/snapshots/cst_output/MultilineComment/same_line/input.sol
