+++ code-analysis/crates/solidity/testing/snapshots/cst_output/VersionPragma/equal_operator/input.sol
