+++ code-analysis/crates/solidity/testing/snapshots/cst_output/VersionPragma/exact_version/input.sol
