+++ code-analysis/crates/solidity/testing/snapshots/cst_output/VersionPragma/less_than_operator/input.sol
