+++ code-analysis/crates/solidity/testing/snapshots/cst_output/VersionPragma/multiple_exact_versions/input.sol
