+++ code-analysis/crates/solidity/testing/snapshots/cst_output/VersionPragma/with_trivia/input.sol
